Modularize Codebase, ci Build implementation, ui Enhancment
