+++ index.html
@@ -3,7 +3,7 @@
   <head>
     <link rel="icon" href="https://treadwell.b-cdn.net/favicon.ico"> 
     <title>Freight Pallet Calculator</title>
-    <link rel="stylesheet" type="text/css" href="style.css" />
+    <link rel="stylesheet" type="text/css" href="src/styles/style.css" />
   </head>
   <body>
     <img
@@ -130,6 +130,6 @@
       <div id="loading" class="loading hidden">Loading.............</div>
       <div id="palletSize" class="pallet-size hidden"></div>
     </div>
-    <script src="src/script.js"></script>
+    <script type="module" src="src/main.js"></script>
   </body>
 </html>
