+++ package.json
@@ -1,5 +1,10 @@
 {
   "dependencies": {
     "vite": "^4.5.0"
+  },
+  "scripts": {
+    "dev": "vite", 
+    "build": "vite build",
+    "preview": "vite preview" 
   }
 }
