--- vite.config.js
@@ -1,9 +1,0 @@
-/** @type {import('vite').UserConfig} */
-export default {
-  // ...{
-  "scripts": {
-    "dev": "vite", // start dev server, aliases: `vite dev`, `vite serve`
-    "build": "vite build", // build for production
-    "preview": "vite preview" // locally preview production build
-  },
-}
